Validate field names to prevent sorting and filtering on invalid fields
This will log warnings to the console for cases like the following:

* Adding a sort/filter for a calculation that must be done in memory
* Adding a sort/filter for a field that does not exist as an attribute, calculation, or aggregate

If an invalid field is provided as a sort/filter from eg. the URL, it
is silently discarded.

Note that we don't account for cases where a query is provided for a Cinder
table that already has invalid sort/filters on it; we only validate what we
want to add to it.

Closes #32